+++ coordinator/src/app_desc/params.rs
@@ -14,44 +14,10 @@
 // You should have received a copy of the GNU General Public License
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 
-use crate::values::Value;
-use crate::{
-    app_desc::{AppDesc, HostSetup, ModuleSetup},
-    values::TOMLValueDeserializer,
-};
+use crate::app_desc::{AppDesc, HostSetup, ModuleSetup};
+use crate::desc_common::params::Merger;
 use anyhow::Context as _;
-use handlebars::{no_escape, Context, Handlebars, TemplateRenderError};
 use std::collections::BTreeMap;
-
-/// You use Handlebars style template in app descriptor.
-///   some-config = "{{other-variable}}"
-/// By starting "@" you can express map or array using template.
-///   some-config-array = "@[ {{variable1}}, {{variable2}} ]"
-///   some-config-map = "@[ \"key\" = {{value-variable}} ]"
-/// If you want to start a string value please use "@@"
-///   some-config = "@@start from @"
-struct Merger<'reg> {
-    registry: Handlebars<'reg>,
-    context: Context,
-}
-
-#[allow(dead_code)]
-impl<'reg> Merger<'reg> {
-    fn new(params: &BTreeMap<String, String>) -> Merger<'reg> {
-        let mut registry = Handlebars::new();
-        registry.register_escape_fn(no_escape);
-        registry.set_strict_mode(true);
-
-        Merger {
-            registry,
-            context: Context::wraps(params).unwrap(),
-        }
-    }
-
-    fn merge(&self, s: &str) -> Result<String, TemplateRenderError> {
-        self.registry.render_template_with_context(s, &self.context)
-    }
-}
 
 impl AppDesc {
     pub fn merge_params(&mut self, params: &BTreeMap<String, String>) -> anyhow::Result<()> {
@@ -92,135 +58,5 @@ impl HostSetup {
             value.merge_params(merger).with_context(|| format!("host > genesis-config > {}", config))?;
         }
         Ok(())
-    }
-}
-
-impl Value {
-    fn merge_params(&mut self, merger: &Merger) -> anyhow::Result<()> {
-        match self {
-            Self::String(s) => {
-                if s.starts_with("@@") {
-                    // Change @@ to @ and type is string
-                    let merged = merger.merge(&s[1..])?;
-                    *self = Value::String(merged);
-                } else if s.starts_with('@') {
-                    // Remove @ and type is anything
-                    let merged = merger.merge(&s[1..])?;
-                    *self = TOMLValueDeserializer::deserialize(&merged)?;
-                } else {
-                    // Type is string
-                    let merged = merger.merge(s)?;
-                    *self = Value::String(merged);
-                }
-            }
-            Self::List(list) => {
-                for v in list {
-                    v.merge_params(merger)?
-                }
-            }
-            Self::Map(map) => {
-                for v in map.values_mut() {
-                    v.merge_params(merger)?
-                }
-            }
-            _ => {}
-        }
-        Ok(())
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::Merger;
-    use crate::values::Value;
-
-    #[test]
-    fn merge_into_string_value() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::String("=={{hello}}==".to_owned());
-        value.merge_params(&merger).unwrap();
-        assert_eq!(value, Value::String("==world==".to_owned()))
-    }
-
-    #[test]
-    fn merge_into_non_string_value() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::Int(123);
-        value.merge_params(&merger).unwrap();
-        assert_eq!(value, Value::Int(123))
-    }
-
-    #[test]
-    fn merge_into_value_in_map() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::Map(
-            vec![(String::from("1"), Value::Int(1)), (String::from("2"), Value::String(String::from("=={{hello}}==")))]
-                .into_iter()
-                .collect(),
-        );
-
-        value.merge_params(&merger).unwrap();
-
-        assert_eq!(
-            value,
-            Value::Map(
-                vec![(String::from("1"), Value::Int(1)), (String::from("2"), Value::String(String::from("==world==")))]
-                    .into_iter()
-                    .collect()
-            )
-        );
-    }
-
-    #[test]
-    fn merge_into_value_in_list() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::List(vec![Value::Int(1), Value::String(String::from("=={{hello}}=="))]);
-
-        value.merge_params(&merger).unwrap();
-
-        assert_eq!(value, Value::List(vec![Value::Int(1), Value::String(String::from("==world=="))]));
-    }
-
-    #[test]
-    fn merge_at() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::String("@\"=={{hello}}==\"".to_owned());
-        value.merge_params(&merger).unwrap();
-        assert_eq!(value, Value::String("==world==".to_owned()))
-    }
-
-    #[test]
-    fn merge_at_array() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::String("@[\"{{hello}}\", \"{{hello}}\"]".to_owned());
-        value.merge_params(&merger).unwrap();
-        assert_eq!(value, Value::List(vec![Value::String("world".to_owned()), Value::String("world".to_owned())]))
-    }
-
-    #[test]
-    fn merge_at_map() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::String("@{ \"key\" = \"{{hello}}\" }".to_owned());
-        value.merge_params(&merger).unwrap();
-        assert_eq!(
-            value,
-            Value::Map(vec![("key".to_string(), Value::String("world".to_string())),].into_iter().collect())
-        )
-    }
-
-    #[test]
-    fn merge_atat() {
-        let params = vec![("hello".to_owned(), "world".to_owned())].into_iter().collect();
-        let merger = Merger::new(&params);
-        let mut value = Value::String("@@=={{hello}}==".to_owned());
-        value.merge_params(&merger).unwrap();
-        assert_eq!(value, Value::String("@==world==".to_owned()))
     }
 }
